requirements.txt
realtime==2.1.0
python-dotenv==1.0.1
pytest-asyncio==0.25.2
pytest==8.3.4
pillow==11.1.0
qrcode==8.0
supabase==2.11.0
pyserial==3.5